+++ libs/actions.refresh-table/src/lib/refresh-table-action-handler.service.spec.ts
@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { TableDataConfiguratorService } from '@spryker/table';
+import {
+  TableDataConfiguratorService,
+  TableLocatorService,
+} from '@spryker/table';
 
 import { RefreshTableActionHandlerService } from './refresh-table-action-handler.service';
 
@@ -11,8 +14,20 @@ const mockActionsConfig = {
 const mockContext = 'mockContext';
 
 @Injectable()
+class MockTable {
+  injector = {
+    get: jest.fn(),
+  };
+}
+
+@Injectable()
 class MockInjector {
   get = jest.fn();
+}
+
+@Injectable()
+class MockTableLocatorService {
+  findById = jest.fn();
 }
 
 @Injectable()
@@ -22,14 +37,22 @@ class MockTableDataConfiguratorService {
 
 describe('RefreshTableActionHandlerService', () => {
   let service: RefreshTableActionHandlerService;
+  let table: MockTable;
   let injector: MockInjector;
+  let tableLocatorService: MockTableLocatorService;
   let tableDataConfiguratorService: MockTableDataConfiguratorService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       providers: [
+        MockTable,
         MockInjector,
+        MockTableLocatorService,
         MockTableDataConfiguratorService,
+        {
+          provide: TableLocatorService,
+          useExisting: MockTableLocatorService,
+        },
         {
           provide: TableDataConfiguratorService,
           useExisting: MockTableDataConfiguratorService,
@@ -38,18 +61,51 @@ describe('RefreshTableActionHandlerService', () => {
     });
 
     service = TestBed.inject(RefreshTableActionHandlerService);
+    table = TestBed.inject(MockTable);
     injector = TestBed.inject(MockInjector);
+    tableLocatorService = TestBed.inject(MockTableLocatorService);
     tableDataConfiguratorService = TestBed.inject(
       MockTableDataConfiguratorService,
     );
 
-    injector.get.mockReturnValue(tableDataConfiguratorService);
+    injector.get.mockImplementation((instance) => {
+      if (instance === TableLocatorService) {
+        return tableLocatorService;
+      }
+
+      if (instance === TableDataConfiguratorService) {
+        return tableDataConfiguratorService;
+      }
+    });
   });
 
   it('should update the table via TableDataConfiguratorService.update()', () => {
     service.handleAction(injector, mockActionsConfig, mockContext);
 
     expect(tableDataConfiguratorService.update).toHaveBeenCalledWith({});
+  });
+
+  it('should find the table by `tableId` via TableLocatorService.findById() and update via TableDataConfiguratorService.update()', () => {
+    tableLocatorService.findById.mockReturnValue(table);
+    table.injector.get.mockReturnValue(tableDataConfiguratorService);
+    service.handleAction(
+      injector,
+      { tableId: 'mockId', ...mockActionsConfig },
+      mockContext,
+    );
+
+    expect(tableLocatorService.findById).toHaveBeenCalledWith('mockId');
+    expect(tableDataConfiguratorService.update).toHaveBeenCalledWith({});
+  });
+
+  it('should throw an error if table is not defined by `tableId`', () => {
+    expect(() =>
+      service.handleAction(
+        injector,
+        { tableId: 'mockId', ...mockActionsConfig },
+        mockContext,
+      ),
+    ).toThrow();
   });
 
   it('should return stream that emits empty value', () => {