+++ libs/actions.refresh-table/src/lib/types.ts
@@ -1,4 +1,5 @@
 import { ActionConfig } from '@spryker/actions';
 
-// tslint:disable-next-line: no-empty-interface
-export interface RefreshTableActionConfig extends ActionConfig {}
+export interface RefreshTableActionConfig extends ActionConfig {
+  tableId?: string;
+}