+++ libs/unsaved-changes.guard.drawer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3-next.0](https://github.com/spryker/ui-components/compare/@spryker/unsaved-changes.guard.drawer@0.1.2...@spryker/unsaved-changes.guard.drawer@0.1.3-next.0) (2021-08-02)
+
+**Note:** Version bump only for package @spryker/unsaved-changes.guard.drawer
+
+
+
+
+
 ## [0.1.2](https://github.com/spryker/ui-components/compare/@spryker/unsaved-changes.guard.drawer@0.1.2-next.0...@spryker/unsaved-changes.guard.drawer@0.1.2) (2021-08-02)
 
 **Note:** Version bump only for package @spryker/unsaved-changes.guard.drawer