+++ libs/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4-next.0](https://github.com/spryker/ui-components/compare/@spryker/utils@0.2.3...@spryker/utils@0.2.4-next.0) (2021-08-02)
+
+**Note:** Version bump only for package @spryker/utils
+
+
+
+
+
 ## [0.2.3](https://github.com/spryker/ui-components/compare/@spryker/utils@0.2.3-next.0...@spryker/utils@0.2.3) (2021-08-02)
 
 **Note:** Version bump only for package @spryker/utils